requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# MQTT
aiomqtt==2.3.0

# AI / Agents
langchain==0.3.14
langchain-openai==0.3.0
langgraph==0.2.60
langchain-community==0.3.14
langchain-chroma==0.2.2

# Storage
chromadb==0.6.3
aiosqlite==0.20.0

# External APIs
httpx==0.28.1
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Voice
elevenlabs==1.50.5

# Utils
pyyaml==6.0.2
python-multipart==0.0.20
websockets==14.1
